backend/requirements.complete.txt
# Smart Parking Platform V6 - Complete Requirements

# ===============================================
# CORE FRAMEWORK
# ===============================================
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-multipart==0.0.9

# ===============================================
# DATABASE
# ===============================================
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.2
psycopg2-binary==2.9.9

# ===============================================
# REDIS
# ===============================================
redis==5.0.8
hiredis==2.4.0

# ===============================================
# AUTHENTICATION & SECURITY
# ===============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.9.0
cryptography==42.0.5

# ===============================================
# CHIRPSTACK INTEGRATION
# ===============================================
grpcio==1.65.5
grpcio-tools==1.65.5
chirpstack-api==4.9.0
protobuf==5.28.2

# ===============================================
# HTTP CLIENTS
# ===============================================
httpx==0.27.2
aiohttp==3.10.5

# ===============================================
# MONITORING & OBSERVABILITY
# ===============================================
prometheus-client==0.20.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0

# ===============================================
# TESTING
# ===============================================
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
hypothesis==6.112.0
locust==2.32.2
faker==28.4.1

# ===============================================
# UTILITIES
# ===============================================
python-dotenv==1.0.1
tenacity==9.0.0
pytz==2024.2
croniter==3.0.3

# ===============================================
# GRAPHQL (OPTIONAL)
# ===============================================
strawberry-graphql[fastapi]==0.242.0

# ===============================================
# DEVELOPMENT
# ===============================================
black==24.8.0
ruff==0.6.8
mypy==1.11.2
ipython==8.27.0
